refactor: split interactive equipment chart
